Stop "filename changed" issue from occuring
This is to address https://github.com/hashicorp/terraform/issues/7613
where lambda filename changes depending on a user that runs Terraform
due to a path on user's local disk being different. This in turn
triggers undesired and annoying lambda funciton updates.
Ignoring filename changes in lifcycle block is not working as expected
- changes to the zip file are correctly picked up due to the sha sum
being different but the path change is not so if user A ran TF
and then user B changes zip contents - terraform will want user A's
path to the file as new path on user B's disk is not passed as it is ignored.