Remove var.lambda_version
As far as I can tell, this variable's purpose was to let people trigger the SNS notification after changing other variables or updating the Lambda function code, but only if they remember to change lambda_version. I have changed it to instead use the Lambda source code hash as a trigger so it automatically sends a notification after anything changes.