Remove var.stack_name and use random id instead
The notification code was expecting stack_name and autoscaling_name to be the same. It seems like it normally would be the same, but then there would be no need for 2 variables. The only reason to have them different would be to have multiple usages of this module against the same autoscaling group.

With the above observations/assumptions, I have removed var.autoscaling_name and var.stack_name and replaced it with var.asg_name, and then used a random id to build an internal stack name to allow for multiple usages.

I have also updated the readme example to remove variables with sensible defaults that should generally not be passed into this module.