Remove stack_name and envname variables
An auto scaling group's name is unique in an AWS account, so we can use asg_name as the sole unique identifier.

The asg_name would usually include the environment name in it, so envname is unnecessary.

This change stops this module from being able to work multiple times for the same ASG. We can't think of when we would want to do that, so that is being dropped in favour of simplicity.